CLI: split context git list/remove/clear
