+++ cmd/prescribe/cmds/context/git/clear.go
@@ -1,0 +1,33 @@
+package git
+
+import (
+	"fmt"
+
+	"github.com/go-go-golems/prescribe/cmd/prescribe/cmds/helpers"
+	"github.com/spf13/cobra"
+)
+
+func newGitContextClearCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "clear",
+		Short: "Clear all git_context items",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			ctrl, err := helpers.NewInitializedController(cmd)
+			if err != nil {
+				return err
+			}
+			helpers.LoadDefaultSessionIfExists(ctrl)
+
+			data := ctrl.GetData()
+			data.GitContext = nil
+
+			savePath := ctrl.GetDefaultSessionPath()
+			if err := ctrl.SaveSession(savePath); err != nil {
+				return err
+			}
+			fmt.Printf("Cleared git_context\n")
+			fmt.Printf("Session saved: %s\n", savePath)
+			return nil
+		},
+	}
+}
