+++ cmd/prescribe/cmds/context/git/legacy.go
@@ -169,103 +169,6 @@ func newGitHistorySetCmd() *cobra.Command {
 	return cmd
 }
 
-func newGitContextListCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "list",
-		Short: "List configured git_context items",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			ctrl, err := helpers.NewInitializedController(cmd)
-			if err != nil {
-				return err
-			}
-			helpers.LoadDefaultSessionIfExists(ctrl)
-
-			items := ctrl.GetData().GitContext
-			if len(items) == 0 {
-				fmt.Println("No git_context items configured")
-				return nil
-			}
-
-			for i, it := range items {
-				switch it.Kind {
-				case domain.GitContextItemKindCommit, domain.GitContextItemKindCommitPatch:
-					fmt.Printf("[%d] %s ref=%s", i, it.Kind, it.Ref)
-					if len(it.Paths) > 0 {
-						fmt.Printf(" paths=%s", strings.Join(it.Paths, ","))
-					}
-					fmt.Printf("\n")
-				case domain.GitContextItemKindFileAtRef:
-					fmt.Printf("[%d] %s ref=%s path=%s\n", i, it.Kind, it.Ref, it.Path)
-				case domain.GitContextItemKindFileDiff:
-					fmt.Printf("[%d] %s from=%s to=%s path=%s\n", i, it.Kind, it.From, it.To, it.Path)
-				default:
-					fmt.Printf("[%d] %s\n", i, it.Kind)
-				}
-			}
-			return nil
-		},
-	}
-}
-
-func newGitContextRemoveCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "remove <index>",
-		Short: "Remove a git_context item by index",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			idx, err := strconv.Atoi(args[0])
-			if err != nil {
-				return fmt.Errorf("invalid index %q", args[0])
-			}
-
-			ctrl, err := helpers.NewInitializedController(cmd)
-			if err != nil {
-				return err
-			}
-			helpers.LoadDefaultSessionIfExists(ctrl)
-
-			data := ctrl.GetData()
-			if idx < 0 || idx >= len(data.GitContext) {
-				return fmt.Errorf("index out of range: %d", idx)
-			}
-
-			data.GitContext = append(data.GitContext[:idx], data.GitContext[idx+1:]...)
-			savePath := ctrl.GetDefaultSessionPath()
-			if err := ctrl.SaveSession(savePath); err != nil {
-				return err
-			}
-			fmt.Printf("Removed git_context item %d\n", idx)
-			fmt.Printf("Session saved: %s\n", savePath)
-			return nil
-		},
-	}
-}
-
-func newGitContextClearCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "clear",
-		Short: "Clear all git_context items",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			ctrl, err := helpers.NewInitializedController(cmd)
-			if err != nil {
-				return err
-			}
-			helpers.LoadDefaultSessionIfExists(ctrl)
-
-			data := ctrl.GetData()
-			data.GitContext = nil
-
-			savePath := ctrl.GetDefaultSessionPath()
-			if err := ctrl.SaveSession(savePath); err != nil {
-				return err
-			}
-			fmt.Printf("Cleared git_context\n")
-			fmt.Printf("Session saved: %s\n", savePath)
-			return nil
-		},
-	}
-}
-
 func newGitContextAddCmd() *cobra.Command {
 	addCmd := &cobra.Command{
 		Use:   "add",
